requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
jinja2==3.1.4
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
apscheduler==3.10.4
lbc==1.1.2
httpx==0.27.2
ovh==1.0.0
python-dotenv==1.0.1
aiosqlite==0.20.0